restruktur url parameter dari `id` menjadi `uuid` dan pemisahan halaman `create` dan `update`
