+++ src/app/(authenticated user)/data/products/(form)/[uuid]/page.tsx
@@ -1,0 +1,12 @@
+'use client'
+
+import { ProductFormPageTemplate } from '../_components/page-template'
+import { useHook } from './_hook'
+
+export default function ProductUpdatePage({
+  params: { uuid },
+}: {
+  params: { uuid: string }
+}) {
+  return <ProductFormPageTemplate {...useHook(uuid)} />
+}
