+++ src/app/(authenticated user)/data/products/(form)/_components/product-form/props.ts
