+++ src/app/(authenticated user)/data/products/(form)/create/page.tsx
@@ -1,0 +1,8 @@
+'use client'
+
+import { ProductFormPageTemplate } from '../_components/page-template'
+import { useHook } from './_hook'
+
+export default function ProductCreatePage() {
+  return <ProductFormPageTemplate {...useHook()} />
+}
